move ip get to realip package, update docs
