+++ realip/real_test.go
@@ -1,4 +1,4 @@
-package rest
+package realip
 
 import (
 	"log"
@@ -11,13 +11,14 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestGetIPAddressFromHeaders(t *testing.T) {
+func TestGetFromHeaders(t *testing.T) {
 	{
 		req, err := http.NewRequest("GET", "/something", http.NoBody)
 		assert.NoError(t, err)
 		req.Header.Add("Something", "1234567")
 		req.Header.Add("X-Real-IP", "8.8.8.8")
-		adr := GetIPAddress(req)
+		adr, err := Get(req)
+		require.NoError(t, err)
 		assert.Equal(t, "8.8.8.8", adr)
 	}
 	{
@@ -25,7 +26,8 @@ func TestGetIPAddressFromHeaders(t *testing.T) {
 		assert.NoError(t, err)
 		req.Header.Add("Something", "1234567")
 		req.Header.Add("X-Forwarded-For", "8.8.8.8,1.1.1.2, 30.30.30.1")
-		adr := GetIPAddress(req)
+		adr, err := Get(req)
+		require.NoError(t, err)
 		assert.Equal(t, "30.30.30.1", adr)
 	}
 	{
@@ -33,7 +35,8 @@ func TestGetIPAddressFromHeaders(t *testing.T) {
 		assert.NoError(t, err)
 		req.Header.Add("Something", "1234567")
 		req.Header.Add("X-Forwarded-For", "8.8.8.8,1.1.1.2,192.168.1.1,10.0.0.65")
-		adr := GetIPAddress(req)
+		adr, err := Get(req)
+		require.NoError(t, err)
 		assert.Equal(t, "1.1.1.2", adr)
 	}
 	{
@@ -42,7 +45,8 @@ func TestGetIPAddressFromHeaders(t *testing.T) {
 		req.Header.Add("Something", "1234567")
 		req.Header.Add("X-Forwarded-For", "30.30.30.1")
 		req.Header.Add("X-Real-Ip", "10.0.0.1")
-		adr := GetIPAddress(req)
+		adr, err := Get(req)
+		require.NoError(t, err)
 		assert.Equal(t, "30.30.30.1", adr)
 	}
 	{
@@ -51,7 +55,8 @@ func TestGetIPAddressFromHeaders(t *testing.T) {
 		req.Header.Add("Something", "1234567")
 		req.Header.Add("X-Forwarded-For", "30.30.30.1")
 		req.Header.Add("X-Real-Ip", "8.8.8.8")
-		adr := GetIPAddress(req)
+		adr, err := Get(req)
+		require.NoError(t, err)
 		assert.Equal(t, "30.30.30.1", adr)
 	}
 	{
@@ -60,46 +65,30 @@ func TestGetIPAddressFromHeaders(t *testing.T) {
 		req.Header.Add("Something", "1234567")
 		req.Header.Add("X-Forwarded-For", "10.0.0.2,192.168.1.1")
 		req.Header.Add("X-Real-Ip", "8.8.8.8")
-		adr := GetIPAddress(req)
+		adr, err := Get(req)
+		require.NoError(t, err)
 		assert.Equal(t, "8.8.8.8", adr)
 	}
 	{
 		req, err := http.NewRequest("GET", "/something", http.NoBody)
 		assert.NoError(t, err)
-		ip := GetIPAddress(req)
+		ip, err := Get(req)
+		assert.Error(t, err)
 		assert.Equal(t, "", ip)
 	}
 }
 
-func TestGetIPAddressFromRemoteAddr(t *testing.T) {
+func TestGetFromRemoteAddr(t *testing.T) {
 	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		log.Printf("%v", r)
-		adr := GetIPAddress(r)
+		adr, err := Get(r)
+		require.NoError(t, err)
 		assert.Equal(t, "127.0.0.1", adr)
 	}))
 
 	req, err := http.NewRequest("GET", ts.URL+"/something", http.NoBody)
 	require.NoError(t, err)
 	client := http.Client{Timeout: time.Second}
-	_, err = client.Do(req)
-	require.NoError(t, err)
-}
-
-func TestRealIP(t *testing.T) {
-
-	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		log.Printf("%v", r)
-		require.Equal(t, "1.2.3.4", r.RemoteAddr)
-		adr := GetIPAddress(r)
-		assert.Equal(t, "1.2.3.4", adr)
-	})
-
-	ts := httptest.NewServer(RealIP(handler))
-
-	req, err := http.NewRequest("GET", ts.URL+"/something", http.NoBody)
-	require.NoError(t, err)
-	client := http.Client{Timeout: time.Second}
-	req.Header.Add("X-Real-IP", "1.2.3.4")
 	_, err = client.Do(req)
 	require.NoError(t, err)
 }
